dev_requirements.txt
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
matplotlib==3.8.3
MedPy==0.4.0
nibabel==5.2.0
numpy==1.26.4
scikit-image==0.22.0
scipy==1.10.1
SimpleITK==2.3.1
tensorboardX==2.6.2.2
pandas==2.2.0
pillow==10.2.0
tqdm
opencv-python
h5py
timm
